Move marpUtils and contants to utils folder
