+++ libs/application/templates/social-insurance-administration/medical-and-rehabilitation-payments/src/fields/Review/review-groups/BaseInformation.tsx
@@ -7,7 +7,7 @@ import { formatNumber } from 'libphonenumber-js'
 import {
   getApplicationAnswers,
   getApplicationExternalData,
-} from '../../../lib/medicalAndRehabilitationPaymentsUtils'
+} from '../../../utils/medicalAndRehabilitationPaymentsUtils'
 import { ReviewGroupProps } from './props'
 
 export const BaseInformation = ({
