+++ libs/application/templates/social-insurance-administration/medical-and-rehabilitation-payments/src/fields/Review/review-groups/Questions.tsx
@@ -6,7 +6,7 @@ import {
 } from '@island.is/application/ui-components'
 import { GridColumn, GridRow, Stack } from '@island.is/island-ui/core'
 import { useLocale } from '@island.is/localization'
-import { getApplicationAnswers } from '../../../lib/medicalAndRehabilitationPaymentsUtils'
+import { getApplicationAnswers } from '../../../utils/medicalAndRehabilitationPaymentsUtils'
 import { medicalAndRehabilitationPaymentsFormMessage } from '../../../lib/messages'
 import { ReviewGroupProps } from './props'
